refactor. LottoTicket클래스 메서드 분리
